Refactor to make it proper module. Separate into two files, where one is the interface and the other provides the convenience functions and classes. Improve handling of invalid queries. Add tests.
